--- src/main/java/com/luckyvicky/woosan/domain/Likes/controller/LikesController.java
@@ -1,4 +1,0 @@
-package com.luckyvicky.woosan.domain.Likes.controller;
-
-public class LikesController {
-}
