--- src/main/java/com/luckyvicky/woosan/domain/Likes/dto/LikesDTO.java
@@ -1,4 +1,0 @@
-package com.luckyvicky.woosan.domain.Likes.dto;
-
-public class LikesDTO {
-}
