--- src/main/java/com/luckyvicky/woosan/domain/Likes/entity/Likes.java
@@ -1,4 +1,0 @@
-package com.luckyvicky.woosan.domain.Likes.entity;
-
-public class Likes {
-}
