--- src/main/java/com/luckyvicky/woosan/domain/Likes/repository/LikesRepository.java
@@ -1,4 +1,0 @@
-package com.luckyvicky.woosan.domain.Likes.repository;
-
-public interface LikesRepository {
-}
