--- src/main/java/com/luckyvicky/woosan/domain/Likes/service/LikesService.java
@@ -1,4 +1,0 @@
-package com.luckyvicky.woosan.domain.Likes.service;
-
-public interface LikesService {
-}
