+++ src/main/java/com/luckyvicky/woosan/domain/likes/controller/LikesController.java
@@ -1,0 +1,4 @@
+package com.luckyvicky.woosan.domain.likes.controller;
+
+public class LikesController {
+}
