+++ src/main/java/com/luckyvicky/woosan/domain/likes/dto/LikesDTO.java
@@ -1,0 +1,4 @@
+package com.luckyvicky.woosan.domain.likes.dto;
+
+public class LikesDTO {
+}
