+++ src/main/java/com/luckyvicky/woosan/domain/likes/entity/Likes.java
@@ -1,0 +1,4 @@
+package com.luckyvicky.woosan.domain.likes.entity;
+
+public class Likes {
+}
