+++ src/main/java/com/luckyvicky/woosan/domain/likes/repository/LikesRepository.java
@@ -1,0 +1,4 @@
+package com.luckyvicky.woosan.domain.likes.repository;
+
+public interface LikesRepository {
+}
