+++ src/main/java/com/luckyvicky/woosan/domain/likes/service/LikesService.java
@@ -1,0 +1,4 @@
+package com.luckyvicky.woosan.domain.likes.service;
+
+public interface LikesService {
+}
